Final tweaks; add one more test

+++ src/java/org/apache/cassandra/index/sai/utils/RowIdWithMeta.java
@@ -34,7 +34,7 @@ public abstract class RowIdWithMeta
         this.segmentRowId = segmentRowId;
     }
 
-    public final long getSegmentRowId()
+    public final int getSegmentRowId()
     {
         return segmentRowId;
     }
